CASSANDRA-20328 Also copy tools/bin/cassandra.in.sh for tools
After CASSANDRA-20328, sstableloader has been updated to be a separate
tool residing in the tools directory.

Existing dtests using sstableloader rely on make_cassandra_env which
copies bin/cassandra.in.sh to the node's directory. As
tools/bin/sstableloader now relies on tools/bin/cassandra.in.sh, updates
this function to also copy this file to the node's path.